updated some of the jpg management scripts in tools

+++ tools/iphone_transcoder/convert.py
@@ -1,8 +1,11 @@
 import os
 import sys
 import shutil
+import time
 
 from azure.storage import BlobService
+
+skipconvert = False
 
 # conf
 output_folder = "output"
@@ -20,21 +23,21 @@ audio_file_name = output_folder + "/audio.mp3"
 image_file_name = output_folder + "/image_%05d.jpg"
 
 # create folder
-if os.path.exists(output_folder):
-	shutil.rmtree(output_folder)
-os.makedirs(output_folder)
+if not skipconvert:
+	if os.path.exists(output_folder):
+		shutil.rmtree(output_folder)
+	os.makedirs(output_folder)
 
-# extract audio
-print "=== Converting Audio ==="
-command = "ffmpeg -i {0} -b:a 192k -map a {1}".format(filename, audio_file_name) 
-os.system(command)
+	# extract audio
+	print "=== Converting Audio ==="
+	command = "ffmpeg -i {0} -b:a 192k -map a {1}".format(filename, audio_file_name) 
+	os.system(command)
 
-print "=== Converting Images ==="
-command = "ffmpeg -i {0} -r 25 -vf scale=640:-1 -q:v 9 -an -f image2 {1}".format(filename, image_file_name) 
-os.system(command)
+	print "=== Converting Images ==="
+	command = "ffmpeg -i {0} -r 25 -vf scale=640:-1 -q:v 9 -an -f image2 {1}".format(filename, image_file_name) 
+	os.system(command)
 
-# if there is no video key
-# return
+# if there is no video key return
 if len(sys.argv) <= 2:
 	exit()
 
@@ -44,9 +47,11 @@ video_key = sys.argv[2]
 
 # walk all files in dir and push to bucket
 key = raw_input("Please enter azure vidoepath blob storage key: ")
-blob_service = BlobService(account_name='videopath', account_key=key)
+blob_service = BlobService(account_name='videopathmobilefiles', account_key=key)
 basepath = os.path.dirname(os.path.abspath(__file__)) + "/" + output_folder
+container_name = video_key.lower()
 print basepath
+
 for path, subdirs, files in os.walk(basepath):
 	for name in files:
 		# don't upload hidden files
@@ -54,9 +59,9 @@ for path, subdirs, files in os.walk(basepath):
 		    continue
 
 		pathname = os.path.join(path, name)
-		keyname = video_key + pathname.replace(basepath, "")
+		keyname = pathname.replace(basepath, "")[1:]
 
-		print pathname + " --> " + keyname
+		print pathname + " --> " + container_name + ": " + keyname
 
 		fileName, fileExtension = os.path.splitext(pathname)
 		content_type = ''
@@ -67,12 +72,16 @@ for path, subdirs, files in os.walk(basepath):
 			content_type = "image/jpeg"
 		elif fileExtension == ".png":
 			content_type = "image/png"
-
-		blob_service.put_block_blob_from_path(
-		    'jpgs',
-		    keyname,
-		    pathname,
-		    x_ms_blob_content_type=content_type,
-		    x_ms_blob_cache_control='public, max-age=600'
-		)
-
+		while True:
+			try:
+				blob_service.put_block_blob_from_path(
+				    container_name,
+				    keyname,
+				    pathname,
+				    x_ms_blob_content_type=content_type,
+				    x_ms_blob_cache_control='public, max-age=600'
+				)
+				break
+			except:
+				print "retrying..."
+				time.sleep(1)
